+++ CHANGES.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented here. The format is based
 on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- *Breaking*: `ref` support, thanks to [`React.forwardRef`](https://reactjs.org/docs/forwarding-refs.html)
+
 ## 2019/03/28 0.6.1
 
 ### Changed